+++ src/lib.rs
@@ -21,18 +21,30 @@ use core::{cmp, ops::RangeBounds};
 /// A map implementation based on skiplist
 pub mod map;
 
+/// Options for the [`SkipMap`](crate::SkipMap).
+pub mod options;
+pub use options::Options;
+#[cfg(all(feature = "memmap", not(target_family = "wasm")))]
+pub use options::{MmapOptions, OpenOptions};
+
 mod types;
 pub use types::*;
 
-#[cfg(all(feature = "memmap", not(target_family = "wasm")))]
-#[cfg_attr(docsrs, doc(cfg(all(feature = "memmap", not(target_family = "wasm")))))]
-pub use rarena_allocator::{MmapOptions, OpenOptions};
-
-pub use rarena_allocator::{Arena, ArenaOptions, Error};
+pub use rarena_allocator::{Arena, Error as ArenaError};
 
 #[cfg(all(feature = "memmap", not(target_family = "wasm")))]
 fn invalid_data<E: std::error::Error + Send + Sync + 'static>(e: E) -> std::io::Error {
   std::io::Error::new(std::io::ErrorKind::InvalidData, e)
+}
+
+#[cfg(all(feature = "memmap", not(target_family = "wasm")))]
+fn bad_magic_version() -> std::io::Error {
+  std::io::Error::new(std::io::ErrorKind::InvalidData, "bad magic version")
+}
+
+#[cfg(all(feature = "memmap", not(target_family = "wasm")))]
+fn bad_version() -> std::io::Error {
+  std::io::Error::new(std::io::ErrorKind::InvalidData, "bad version")
 }
 
 pub use map::{AllVersionsIter, SkipMap};